🚚 Rename convert.js to convertLodashFp.js
