+++ packages/immutadot-lodash/src/array/pullAt.js
@@ -1,5 +1,5 @@
 import _pullAt from 'lodash/fp/pullAt'
-import { convertLodashFp } from 'util/convert'
+import { convertLodashFp } from 'util/convertLodashFp'
 
 /**
  * Replaces an array removing the specified indexes from the former array.
