+++ packages/immutadot-lodash/src/object/defaults.js
@@ -1,5 +1,5 @@
 import _defaults from 'lodash/fp/defaults'
-import { convertLodashFp } from 'util/convert'
+import { convertLodashFp } from 'util/convertLodashFp'
 
 /**
  * Replaces by an object assigning own and inherited enumerable string keyed properties of source objects to the destination object for all destination properties that resolve to <code>undefined</code>.<br >
